fix: video info bizId type

+++ src/main/kotlin/moe/kurenai/bot/model/bilibili/VideoInfo.kt
@@ -61,7 +61,7 @@ data class VideoInfo(
             val rawText: String,
             val type: Int,
             @SerialName("biz_id")
-            val bizId: Int
+            val bizId: Long
         )
 
         @Serializable
